add snapshot and write
