+++ new-mock-engine-store/src/mock_store/mock_fast_add_peer_impls.rs
