+++ new-mock-engine-store/src/mock_store/mod.rs
@@ -3,7 +3,9 @@ pub(crate) mod common;
 pub mod mock_core;
 pub use mock_core::*;
 pub mod mock_engine_store_server;
-pub mod mock_fast_add_peer;
+pub(crate) mod mock_fast_add_peer_impls;
+pub(crate) mod mock_snapshot_impls;
+pub(crate) mod mock_write_impls;
 pub use mock_engine_store_server::*;
 pub mod mock_page_storage;
 pub use mock_page_storage::*;
